Pizzeria createOrder test

+++ Java_Fundamentals/OOP/Pizzeria/src/main/java/Models/Pizza.java
@@ -4,17 +4,17 @@ import java.util.List;
 import java.util.Map;
 
 public class Pizza {
-    private final int price;
+    private final double price;
     private final String name;
     private final Map<String,Integer> ingredientsNeeded;
 
-    public Pizza(int price, String name, Map<String, Integer> ingredientsNeeded) {
+    public Pizza(double price, String name, Map<String, Integer> ingredientsNeeded) {
         this.price = price;
         this.name = name;
         this.ingredientsNeeded = ingredientsNeeded;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
